:art: (606): refacto platform infos & make meeting fields optionnal
